Rename `jax_utils` to `conversion`
