+++ tests/python_compiler/dialects/test_transform_dialect.py
@@ -33,9 +33,9 @@ from xdsl.dialects.transform import AnyOpType
 from xdsl.utils.exceptions import VerifyException
 from xdsl.utils.test_value import create_ssa_value
 
+from pennylane.compiler.python_compiler.conversion import xdsl_from_docstring
 from pennylane.compiler.python_compiler.dialects import transform
 from pennylane.compiler.python_compiler.dialects.transform import ApplyRegisteredPassOp
-from pennylane.compiler.python_compiler.jax_utils import xdsl_from_docstring
 from pennylane.compiler.python_compiler.pass_api import (
     ApplyTransformSequence,
     compiler_transform,
