+++ tests/python_compiler/test_python_compiler.py
@@ -38,12 +38,12 @@ from xdsl.interpreters import Interpreter
 import pennylane as qml
 from pennylane.capture import enabled as capture_enabled
 from pennylane.compiler.python_compiler import Compiler
-from pennylane.compiler.python_compiler.dialects import transform
-from pennylane.compiler.python_compiler.jax_utils import (
+from pennylane.compiler.python_compiler.conversion import (
     jax_from_docstring,
     module,
     xdsl_from_docstring,
 )
+from pennylane.compiler.python_compiler.dialects import transform
 from pennylane.compiler.python_compiler.pass_api import (
     ApplyTransformSequence,
     TransformFunctionsExt,
